+++ src/printer_ast.ml
@@ -878,8 +878,9 @@ let pp_specification fmt (v : lident specification) =
       pp_pterm d.body
   in
   let pp_fail fmt (f : lident fail) =
-    Format.fprintf fmt "%a with (%a : %a):@\n  @[%a@];@\n"
+    Format.fprintf fmt "%a with %a(%a : %a):@\n  @[%a@];@\n"
       pp_id f.label
+      (pp_option pp_id) f.fid
       pp_id f.arg
       pp_type f.atype
       pp_pterm f.formula